% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databaseConnection.R
\name{databaseConnection}
\alias{databaseConnection}
\title{To connect to a MySQL database}
\usage{
databaseConnection(input, output, session)
}
\arguments{
\item{host}{Where the server is running. The default value is \emph{localhost}.}

\item{port}{A port number to use for the connection, for connections made using \strong{TCP/IP}.}

\item{dbname}{database name}

\item{user}{A mysql account grant privileges to manipulate table}

\item{password}{for security purposes.}
}
\description{
To connect to a MySQL database
}

