% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadData}
\alias{loadData}
\title{Retrieve function that simply load responses from MySQL.}
\usage{
loadData(input, output, session, table)
}
\description{
Retrieve function that simply load responses from MySQL.
}

