% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDataById.R
\name{loadDataById}
\alias{loadDataById}
\title{Retrieve function that simply load route from MySQL.}
\usage{
loadDataById(input, output, session, id, table)
}
\description{
Retrieve function that simply load route from MySQL.
}

