% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{loadDataFactors}
\alias{loadDataFactors}
\title{Load function that stores factors for route in database.}
\usage{
loadDataFactors(id)
}
\arguments{
\item{...}{loadDataFactors(id)}
}
\description{
Load function that stores factors for route in database.
}

