% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{saveData}
\alias{saveData}
\title{Save function that simply stores responses in MySQL.}
\usage{
saveData(data)
}
\description{
Save function that simply stores responses in MySQL.
}
\examples{
saveData(data)
}

