% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{saveFactorsData}
\alias{saveFactorsData}
\title{Save function that stores factors forr a specific route in database.}
\usage{
saveFactorsData(data)
}
\arguments{
\item{destination}{saveFactorsData(data)}
}
\description{
Save function that stores factors forr a specific route in database.
}

