% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{saveRouteData}
\alias{saveRouteData}
\title{Save function that stores route in database.}
\usage{
saveRouteData(data)
}
\arguments{
\item{destination}{saveRouteData(data)}
}
\description{
Save function that stores route in database.
}

