% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{loadDataFactors}
\alias{loadDataFactors}
\title{Load function that stores factors for route in database.}
\usage{
loadDataFactors(id)
}
\arguments{
\item{id}{route id from saving factors.
loadDataFactors(id)}

\item{tableFactors}{table name for factors.}
}
\description{
Load function that stores factors for route in database.
}

