% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveData.R
\name{saveData}
\alias{saveData}
\title{Save function that simply stores responses in MySQL.}
\usage{
saveData(input, output, session, data, table)
}
\arguments{
\item{data}{inputs from user.}
}
\description{
Save function that simply stores responses in MySQL.
}
\examples{
saveData(data)
}

